move CpgLoader to cpgloading package
