+++ cpgqueryingtests/src/test/scala/io/shiftleft/cpgqueryingtests/codepropertygraph/CpgFactory.scala
@@ -3,7 +3,7 @@ package io.shiftleft.cpgqueryingtests.codepropertygraph
 import java.io.{File, PrintWriter}
 import java.nio.file.Files
 
-import io.shiftleft.CpgLoader
+import io.shiftleft.cpgloading.CpgLoader
 import io.shiftleft.queryprimitives.steps.starters.Cpg
 
 class CpgFactory(frontend: LanguageFrontend) {
